[BSE-4341] Update benchmarking and finalize scripts (#31)
* refactored nyc_taxi benchmark related files -> nyc_taxi directory to match the structure of the tpch benchmark and set us up for adding other benchmarks.
* Added script to run bodo example on platform using SDK, currently supports "WORKSPACE" type although could be moved to github once we fix that API
* Renamed files with typo in name
* Other cleanup to have dask, bodo example write to parquet file on one of the workers.
