+++ benchmarks/nyc_taxi/modin-ray/modin-cluster.yaml
@@ -1,4 +1,4 @@
-# Copied from https://github.com/modin-project/modin/blob/main/examples/tutorial/jupyter/execution/pandas_on_ray/cluster/modin-cluster.yaml
+# Copied from https://github.com/modin-project/modin/blob/main/examples/tutorial/jupyter/execution/pandas_on_ray/cluster/modin-cluster.yaml with some modifications.
 
 # An unique identifier for the head node and workers of this cluster.
 cluster_name: modin_init
@@ -34,7 +34,7 @@ docker:
     # worker_run_options: []
 
 # If a node is idle for this many minutes, it will be removed.
-idle_timeout_minutes: 5
+idle_timeout_minutes: 60
 
 # Cloud-provider specific configuration.
 provider:
@@ -44,7 +44,7 @@ provider:
     # Availability zone(s), comma-separated, that nodes may be launched in.
     # Nodes will be launched in the first listed availability zone and will
     # be tried in the subsequent availability zones if launching fails.
-    availability_zone: us-east-2a,us-east-2b
+    availability_zone: us-east-2c, us-east-2b, us-east-2a
     # Whether to allow node reuse. If set to False, nodes will be terminated
     # instead of stopped.
     cache_stopped_nodes: False # If not present, the default is True.
@@ -73,7 +73,7 @@ available_node_types:
         # For more documentation on available fields, see:
         # http://boto3.readthedocs.io/en/latest/reference/services/ec2.html#EC2.ServiceResource.create_instances
         node_config:
-            InstanceType: r6i.8xlarge
+            InstanceType: r6i.16xlarge
             # Default AMI for us-east-2.
             # Check https://github.com/ray-project/ray/blob/master/python/ray/autoscaler/_private/aws/config.py
             # for default images for other zones.
@@ -82,13 +82,13 @@ available_node_types:
             BlockDeviceMappings:
                 - DeviceName: /dev/sda1
                   Ebs:
-                      VolumeSize: 500
+                      VolumeSize: 1000
                       VolumeType: gp3
             # Additional options in the boto docs.
     ray.worker.default:
         # The minimum number of worker nodes of this type to launch.
         # This number should be >= 0.
-        min_workers: 3
+        min_workers: 1
         # The maximum number of worker nodes of this type to launch.
         # This takes precedence over min_workers.
         max_workers: 3
@@ -103,7 +103,7 @@ available_node_types:
         # For more documentation on available fields, see:
         # http://boto3.readthedocs.io/en/latest/reference/services/ec2.html#EC2.ServiceResource.create_instances
         node_config:
-            InstanceType: r6i.8xlarge
+            InstanceType: r6i.16xlarge
             # Default AMI for us-east-2.
             # Check https://github.com/ray-project/ray/blob/master/python/ray/autoscaler/_private/aws/config.py
             # for default images for other zones.
@@ -112,7 +112,7 @@ available_node_types:
             BlockDeviceMappings:
                 - DeviceName: /dev/sda1
                   Ebs:
-                      VolumeSize: 500
+                      VolumeSize: 1000
                       VolumeType: gp3
             # Run workers on spot by default. Comment this out to use on-demand.
             # NOTE: If relying on spot instances, it is best to specify multiple different instance
