+++ benchmarks/nyc_taxi/modin-ray/nyc_taxi_precipitation.py
