+++ benchmarks/nyc_taxi/spark/spark_nyc_taxi_precipitation.py
