+++ benchmarks/nyc_taxi/spark/wait_for_step.sh
